fix: move genkey to internal
