+++ internal/genkey/types.go
