+++ internal/genkey/validate.go
